LPD-101547 Move the language client extension test to Language Override
The test covers the language client extension, which is built into a batch
engine payload of com.liferay.portal.language.rest.dto.v1_0.Message and
imported as language overrides. That behavior belongs to portal-language, not
to client-extension-web, which only owns the Client Extensions admin UI, so the
test now runs under portal-language-override-web and reports to the Language
Override component.

It needs the sample workspace language client extension deployed and a
locales.enabled list wider than the portal default, so it gets its own project
rather than an env folder on portal-language-override-web.main.

The expected ja-JP translation was also stale since LPS-77699 refreshed the
sample workspace translations.
