+++ modules/apps/portal-language/portal-language-override-web/test.properties
@@ -1,6 +1,7 @@
 modified.files.includes[relevant][portal-language-override-web-playwright-rule]=**
 
 playwright.projects.includes[playwright-js-tomcat101-postgresql163][relevant][portal-language-override-web-playwright-rule]=\
+    portal-language-override-web.client-extension,\
     portal-language-override-web.main
 
 relevant.rule.names=portal-language-override-web-playwright-rule
