+++ modules/apps/portal-language/test.properties
@@ -7,6 +7,7 @@ modules.includes.required.test.batch.class.names.includes[modules-unit][relevant
     apps/portal-language/**/*Test.java
 
 playwright.projects.includes[playwright-js-tomcat101-postgresql163][relevant]=\
+    portal-language-override-web.client-extension,\
     portal-language-override-web.main
 
 relevant.rule.names=portal-language-rule
