+++ modules/test/playwright/tests/portal-language-override-web/client-extension/test.properties
@@ -1,0 +1,1 @@
+testray.main.component.name=Language Override
