+++ test.properties
@@ -5094,6 +5094,7 @@
         frontend-js-item-selector-web.main,\
         frontend-taglib-clay.main,\
         frontend-theme.main,\
+        portal-language-override-web.client-extension,\
         portal-language-override-web.main,\
         product-navigation-applications-menu.main,\
         product-navigation-control-menu-web.main,\
